+++ solat_cb/mle/solve.py
@@ -5,483 +5,11 @@ import os
 import pickle as pl
 from solat_cb.spectra import Spectra
 from solat_cb.simulation import CMB, LATsky
-from solat_cb import mpi
+from solat_cb.mle.utils import *
+from solat_cb.mle.ls import LinearSystem
 
 rad2arcmin = 180*60/np.pi
-
-def effective_ell(lmax, bin_conf):
-    (ib_grid, il_grid, w_array, ell_array) = bin_conf
-    ell = np.arange(0, lmax+1, 1)
-    return np.sum(w_array[ib_grid,il_grid]*ell[ell_array[ib_grid,il_grid]], axis=1)
-
-def bin_from_edges(start, end):
-    nls  = np.amax(end)
-    lmax = nls-1
-    ells, bpws, weights = [], [], []
-    for ib, (li, le) in enumerate(zip(start, end)):
-        nlb      = int(le - li)
-        ells    += list(range(li, le))
-        bpws    += [ib] * nlb
-        weights += [1./nlb] * nlb
-    ells = np.array(ells); bpws = np.array(bpws); weights = np.array(weights)
-
-    nell       = len(ells)
-    n_bands    = bpws[-1]+1
-    nell_array = np.zeros(n_bands, dtype=int)
-    for ii in range(0, nell, 1):
-        if ells[ii]<=lmax and bpws[ii]>=0:
-            nell_array[bpws[ii]]+=1   
-            
-    ell_list, w_list = [], []
-    for ii in range(0, n_bands, 1):
-        ell_list.append(np.zeros(nell_array[ii], dtype=int))
-        w_list.append(np.zeros(nell_array[ii], dtype=np.float64))
-    
-    nell_array *= 0
-    for ii in range(0, nell, 1):
-        if ells[ii]<=lmax and bpws[ii]>=0: 
-            ell_list[bpws[ii]][nell_array[bpws[ii]]] = ells[ii]
-            w_list[bpws[ii]][nell_array[bpws[ii]]]   = weights[ii]
-            nell_array[bpws[ii]]+=1
-
-    for ii in range(0, n_bands, 1):
-        norm=0
-        for jj in range(0, nell_array[ii], 1):
-            norm += w_list[ii][jj]
-        if norm<=0:
-            print(f"Weights in band {ii} are wrong\n")
-        for jj in range(0, nell_array[ii], 1):
-            w_list[ii][jj]/=norm
-
-    return (n_bands, nell_array, ell_list, w_list)
-
-def bin_configuration(info):
-    (n_bands, nell_array, ell_list, w_list) = info
-    ib_array          = np.arange(0, n_bands,       1, dtype=int)
-    il_array          = np.arange(0, nell_array[0], 1, dtype=int)
-    (ib_grid,il_grid) = np.meshgrid(ib_array, il_array, indexing='ij')
-    return (ib_grid, il_grid, np.array(w_list), np.array(ell_list))
-
-#TODO might change the shape of the output array when optimising linear system terms
-def bin_spec_matrix(spec, info):
-    (ib_grid, il_grid, w_array, ell_array) = info
-    return np.sum(w_array[ib_grid,il_grid]*spec[:,:,ell_array[ib_grid,il_grid]], axis=3)
-
-#TODO might change the shape of the output array when optimising linear system terms
-def bin_cov_matrix(cov, info):
-    (ib_grid, il_grid, w_array, ell_array) = info
-    return np.sum(w_array[ib_grid,il_grid]**2*cov[:,:,:,ell_array[ib_grid,il_grid]], axis=4)
-
-def moving_sum(a, n) :
-    ret = np.cumsum(a, dtype=float)
-    ret[n:] = ret[n:] - ret[:-n]
-    return ret[n - 1:] 
-
-
-class Result:
-    
-    def __init__(self, spec, fit, sim, 
-                 alpha_per_split, rm_same_tube,
-                 binwidth, bmin, bmax,
-                 beta_ini=0.0, alpha_ini=0.0):
-
-        # information I should record to remember how the result was calculated
-        # information about the spectra
-        self.specdir  = spec.lat.libdir
-        self.temp_bp  = spec.temp_bp
-        self.aposcale = spec.aposcale
-        self.CO       = spec.CO
-        self.PS       = spec.PS
-        self.pureB    = spec.pureB
-        self.nside    = spec.nside
-        self.fsky     = spec.fsky
-        self.sim_idx  = sim 
-        # information about the fit
-        self.nlb             = binwidth
-        self.bmin            = bmin
-        self.bmax            = bmax
-        self.fit             = fit
-        self.rm_same_tube    = rm_same_tube
-        self.alpha_per_split = alpha_per_split
-        # parameters to calculate
-        self.ml         = {}
-        self.std_fisher = {}
-        self.cov_fisher = {"Iter 0":None }
-        self.variables  = ''
-        if fit=="alpha":
-            ext_par = 0
-            # add them later once you know how many are there
-            self.ml["Iter 0"]         = {}
-            self.std_fisher["Iter 0"] = {}
-        elif fit=="Ad + alpha":
-            ext_par = 1
-            self.ml["Iter 0"]         = { 'Ad':1.0  } 
-            self.std_fisher["Iter 0"] = { 'Ad':None }
-            self.variables           += 'Ad'
-        elif fit=="beta + alpha":
-            ext_par = 1
-            self.ml["Iter 0"]         = { 'beta':beta_ini  }
-            self.std_fisher["Iter 0"] = { 'beta':None }
-            self.variables           += 'beta'
-        elif fit=="As + Ad + alpha":
-            ext_par = 2
-            self.ml["Iter 0"]         = { 'As':1.0,  'Ad':1.0  }
-            self.std_fisher["Iter 0"] = { 'As':None, 'Ad':None }
-            self.variables           += 'As, Ad'
-        elif fit=="Ad + beta + alpha":
-            ext_par = 2
-            self.ml["Iter 0"]         = { 'Ad':1.0,  'beta':beta_ini}
-            self.std_fisher["Iter 0"] = { 'Ad':None, 'beta':0.0 }
-            self.variables           += 'Ad, beta'
-        elif fit=="As + Ad + beta + alpha":
-            ext_par = 3
-            self.ml["Iter 0"]         = { 'As':1.0, 'Ad':1.0, 'beta':beta_ini}
-            self.std_fisher["Iter 0"] = { 'As':None,'Ad':None,'beta':None }
-            self.variables           += 'As, Ad, beta'
-        elif fit=="As + Asd + Ad + alpha":
-            ext_par = 3
-            self.ml["Iter 0"]         = { 'As':1.0,  'Asd':1.0,  'Ad':1.0 }
-            self.std_fisher["Iter 0"] = { 'As':None, 'Asd':None, 'Ad':None}
-            self.variables           += 'As, Asd, Ad' 
-        elif fit=="As + Asd + Ad + beta + alpha":
-            ext_par = 4
-            self.ml["Iter 0"]         = { 'As':1.0, 'Asd':1.0,  'Ad':1.0, 'beta':beta_ini}
-            self.std_fisher["Iter 0"] = { 'As':None,'Asd':None, 'Ad':None,'beta':None }
-            self.variables           += 'As, Asd, Ad, beta'
-            
-        if alpha_per_split:
-            self.Nalpha = spec.Nbands
-            for ii, band in enumerate(spec.bands):
-                self.ml["Iter 0"][band]         = alpha_ini
-                self.std_fisher["Iter 0"][band] = None
-                self.variables                 += f'{band}'if (ii==0 and fit=="alpha") else f', {band}'
-        else:
-            self.Nalpha = spec.Nfreq
-            for ii, freq in enumerate(spec.freqs):
-                self.ml["Iter 0"][freq]         = alpha_ini
-                self.std_fisher["Iter 0"][freq] = None
-                self.variables                 += f'{freq}'if (ii==0 and fit=="alpha") else f', {freq}'
-
-        self.ext_par = ext_par
-        self.Nvar    = self.Nalpha + self.ext_par
-        
-      
-class LinearSystem:
-    mode_options = ['total', 'cumulative', 'ell']
-
-    def __init__(self, mle, inv_cov, mode="total", window=5):
-        self.dt             = np.float64
-        self.fit            = mle.fit
-        self.bin_cl         = mle.bin_terms
-        self.iC             = inv_cov
-        self.Nbands         = mle.Nbands
-        self.mle            = mle
-        
-        assert mode in self.mode_options, f"mode must be one of {self.mode_options}"
-        self.mode           = mode
-        self.window         = window
-        self.Nbins          = mle.Nbins
-        if mode=="total":
-            self.ext_dim = 1
-        elif mode=="cumulative":
-            self.ext_dim = self.Nbins
-        elif mode=="ell":
-            self.ext_dim = self.Nbins - (self.window - 1)
-
-        # common to all fits (basic alpha*alpha fit)
-        self.B_ijpq         = np.zeros((self.Nbands, self.Nbands, self.Nbands, self.Nbands, self.ext_dim), dtype=self.dt)
-        self.E_ijpq         = np.zeros((self.Nbands, self.Nbands, self.Nbands, self.Nbands, self.ext_dim), dtype=self.dt)
-        self.I_ijpq         = np.zeros((self.Nbands, self.Nbands, self.Nbands, self.Nbands, self.ext_dim), dtype=self.dt)
-        self.D_ij           = np.zeros((self.Nbands, self.Nbands, self.ext_dim), dtype=self.dt)
-        self.H_ij           = np.zeros((self.Nbands, self.Nbands, self.ext_dim), dtype=self.dt)
-        self.A              = np.zeros(self.ext_dim, dtype=self.dt)
-        # only used in particular combinations
-        if 'beta' in self.fit:
-            # beta*beta and beta*alpha 
-            self.tau_ij     = np.zeros((self.Nbands, self.Nbands, self.ext_dim), dtype=self.dt)
-            self.varphi_ij  = np.zeros((self.Nbands, self.Nbands, self.ext_dim), dtype=self.dt)
-            self.ene_ij     = np.zeros((self.Nbands, self.Nbands, self.ext_dim), dtype=self.dt) 
-            self.epsilon_ij = np.zeros((self.Nbands, self.Nbands, self.ext_dim), dtype=self.dt)
-            self.C          = np.zeros(self.ext_dim, dtype=self.dt) 
-            self.F          = np.zeros(self.ext_dim, dtype=self.dt) 
-            self.G          = np.zeros(self.ext_dim, dtype=self.dt) 
-            self.O          = np.zeros(self.ext_dim, dtype=self.dt) 
-            self.P          = np.zeros(self.ext_dim, dtype=self.dt)
-        if 'Ad' in self.fit:
-            # Ad*Ad and Ad*alpha
-            self.sigma_ij   = np.zeros((self.Nbands, self.Nbands, self.ext_dim), dtype=self.dt)
-            self.omega_ij   = np.zeros((self.Nbands, self.Nbands, self.ext_dim), dtype=self.dt)
-            self.R          = np.zeros(self.ext_dim, dtype=self.dt)
-            self.N          = np.zeros(self.ext_dim, dtype=self.dt)
-            if 'beta' in self.fit:
-                # Ad*beta
-                self.LAMBDA = np.zeros(self.ext_dim, dtype=self.dt)
-                self.mu     = np.zeros(self.ext_dim, dtype=self.dt)
-        if 'As' in self.fit:
-            # As*As and As*alpha
-            self.nu_ij      = np.zeros((self.Nbands, self.Nbands, self.ext_dim), dtype=self.dt)
-            self.psi_ij     = np.zeros((self.Nbands, self.Nbands, self.ext_dim), dtype=self.dt) 
-            self.S          = np.zeros(self.ext_dim, dtype=self.dt)
-            self.J          = np.zeros(self.ext_dim, dtype=self.dt)
-            if 'beta' in self.fit:
-                # As*beta
-                self.X      = np.zeros(self.ext_dim, dtype=self.dt)
-                self.Y      = np.zeros(self.ext_dim, dtype=self.dt)
-            if 'Ad' in self.fit:
-                # As*Ad
-                self.W      = np.zeros(self.ext_dim, dtype=self.dt)
-        if 'Asd' in self.fit:
-            # Asd*Asd and Asd*alpha
-            self.pi_ij      = np.zeros((self.Nbands, self.Nbands, self.ext_dim), dtype=self.dt)
-            self.rho_ij     = np.zeros((self.Nbands, self.Nbands, self.ext_dim), dtype=self.dt)
-            self.phi_ij     = np.zeros((self.Nbands, self.Nbands, self.ext_dim), dtype=self.dt) 
-            self.OMEGA_ij   = np.zeros((self.Nbands, self.Nbands, self.ext_dim), dtype=self.dt)
-            self.T          = np.zeros(self.ext_dim, dtype=self.dt)
-            self.U          = np.zeros(self.ext_dim, dtype=self.dt)
-            self.Z          = np.zeros(self.ext_dim, dtype=self.dt)
-            self.M          = np.zeros(self.ext_dim, dtype=self.dt)
-            self.L          = np.zeros(self.ext_dim, dtype=self.dt)
-            if 'beta' in self.fit:
-                # Asd*beta
-                self.DELTA  = np.zeros(self.ext_dim, dtype=self.dt)
-                self.eta    = np.zeros(self.ext_dim, dtype=self.dt) 
-                self.theta  = np.zeros(self.ext_dim, dtype=self.dt) 
-                self.delta  = np.zeros(self.ext_dim, dtype=self.dt)
-            if 'Ad' in self.fit:
-                # Asd*Ad
-                self.K      = np.zeros(self.ext_dim, dtype=self.dt)
-                self.xi     = np.zeros(self.ext_dim, dtype=self.dt)
-            if 'As' in self.fit:
-                # Asd*As
-                self.Q      = np.zeros(self.ext_dim, dtype=self.dt)
-                self.V      = np.zeros(self.ext_dim, dtype=self.dt)
-
-    def compute_terms(self):
-        if self.mode=="total":
-            return self.__total__()
-        elif self.mode=="cumulative":
-            return self.__cumulative__() 
-        elif self.mode=="ell":
-            return self.__ell__()
-          
-    def __cumulative__(self):
-        for MN_pair in self.mle.MNidx:
-            ii, jj, pp, qq, mm, nn = self.mle.get_index(MN_pair)
-            # common to all fits (basic alpha*alpha fit)
-            self.B_ijpq[ii,jj,pp,qq,:]    = np.cumsum(self.bin_cl['BBo_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['BBo_ij_b'][pp,qq,:])
-            self.E_ijpq[ii,jj,pp,qq,:]    = np.cumsum(self.bin_cl['EEo_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EEo_ij_b'][pp,qq,:])
-            self.I_ijpq[ii,jj,pp,qq,:]    = np.cumsum(self.bin_cl['BBo_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EEo_ij_b'][pp,qq,:]) 
-            self.D_ij[ii,jj,:]           += np.cumsum(self.bin_cl['EEo_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EBo_ij_b'][pp,qq,:])
-            self.H_ij[ii,jj,:]           += np.cumsum(self.bin_cl['BBo_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EBo_ij_b'][pp,qq,:])
-            self.A[:]                    += np.cumsum(self.bin_cl['EBo_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EBo_ij_b'][pp,qq,:])
-            # only used in particular combinations
-            if 'beta' in self.fit:
-                # beta*beta and beta*alpha 
-                self.tau_ij[ii,jj,:]     += np.cumsum(self.bin_cl['EEo_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EEcmb_ij_b'][pp,qq,:])
-                self.varphi_ij[ii,jj,:]  += np.cumsum(self.bin_cl['EEo_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['BBcmb_ij_b'][pp,qq,:])
-                self.ene_ij[ii,jj,:]     += np.cumsum(self.bin_cl['BBo_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EEcmb_ij_b'][pp,qq,:])
-                self.epsilon_ij[ii,jj,:] += np.cumsum(self.bin_cl['BBo_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['BBcmb_ij_b'][pp,qq,:])
-                self.C[:]                += np.cumsum(self.bin_cl['EEcmb_ij_b'][ii,jj,:] *self.iC[:,mm,nn]* self.bin_cl['BBcmb_ij_b'][pp,qq,:]) 
-                self.F[:]                += np.cumsum(self.bin_cl['EEcmb_ij_b'][ii,jj,:] *self.iC[:,mm,nn]* self.bin_cl['EEcmb_ij_b'][pp,qq,:]) 
-                self.G[:]                += np.cumsum(self.bin_cl['BBcmb_ij_b'][ii,jj,:] *self.iC[:,mm,nn]* self.bin_cl['BBcmb_ij_b'][pp,qq,:]) 
-                self.O[:]                += np.cumsum(self.bin_cl['EBo_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EEcmb_ij_b'][pp,qq,:]) 
-                self.P[:]                += np.cumsum(self.bin_cl['EBo_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['BBcmb_ij_b'][pp,qq,:])
-            if 'Ad' in self.fit:
-                # Ad*Ad and Ad*alpha
-                self.sigma_ij[ii,jj,:]   += np.cumsum(self.bin_cl['EEo_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EBd_ij_b'][pp,qq,:])
-                self.omega_ij[ii,jj,:]   += np.cumsum(self.bin_cl['BBo_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EBd_ij_b'][pp,qq,:])
-                self.R[:]                += np.cumsum(self.bin_cl['EBd_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EBd_ij_b'][pp,qq,:])
-                self.N[:]                += np.cumsum(self.bin_cl['EBo_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EBd_ij_b'][pp,qq,:])
-                if 'beta' in self.fit:
-                    # Ad*beta
-                    self.LAMBDA[:]       += np.cumsum(self.bin_cl['EBd_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EEcmb_ij_b'][pp,qq,:])
-                    self.mu[:]           += np.cumsum(self.bin_cl['EBd_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['BBcmb_ij_b'][pp,qq,:])
-            if 'As' in self.fit:
-                # As*As and As*alpha
-                self.nu_ij[ii,jj,:]      += np.cumsum(self.bin_cl['EEo_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EBs_ij_b'][pp,qq,:])
-                self.psi_ij[ii,jj,:]     += np.cumsum(self.bin_cl['BBo_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EBs_ij_b'][pp,qq,:]) 
-                self.S[:]                += np.cumsum(self.bin_cl['EBs_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EBs_ij_b'][pp,qq,:])
-                self.J[:]                += np.cumsum(self.bin_cl['EBo_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EBs_ij_b'][pp,qq,:])
-                if 'beta' in self.fit:
-                    # As*beta
-                    self.X[:]            += np.cumsum(self.bin_cl['EBs_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EEcmb_ij_b'][pp,qq,:])
-                    self.Y[:]            += np.cumsum(self.bin_cl['EBs_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['BBcmb_ij_b'][pp,qq,:])
-                if 'Ad' in self.fit:
-                    # As*Ad
-                    self.W[:]            += np.cumsum(self.bin_cl['EBs_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EBd_ij_b'][pp,qq,:])
-            if 'Asd' in self.fit:
-                # Asd*Asd and Asd*alpha
-                self.pi_ij[ii,jj,:]      += np.cumsum(self.bin_cl['EEo_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EsBd_ij_b'][pp,qq,:])
-                self.rho_ij[ii,jj,:]     += np.cumsum(self.bin_cl['EEo_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EdBs_ij_b'][pp,qq,:])
-                self.phi_ij[ii,jj,:]     += np.cumsum(self.bin_cl['BBo_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EsBd_ij_b'][pp,qq,:])
-                self.OMEGA_ij[ii,jj,:]   += np.cumsum(self.bin_cl['BBo_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EdBs_ij_b'][pp,qq,:])
-                self.T[:]                += np.cumsum(self.bin_cl['EsBd_ij_b'][ii,jj,:]  *self.iC[:,mm,nn]* self.bin_cl['EsBd_ij_b'][pp,qq,:])
-                self.U[:]                += np.cumsum(self.bin_cl['EdBs_ij_b'][ii,jj,:]  *self.iC[:,mm,nn]* self.bin_cl['EdBs_ij_b'][pp,qq,:])
-                self.Z[:]                += np.cumsum(self.bin_cl['EsBd_ij_b'][ii,jj,:]  *self.iC[:,mm,nn]* self.bin_cl['EdBs_ij_b'][pp,qq,:])
-                self.M[:]                += np.cumsum(self.bin_cl['EBo_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EdBs_ij_b'][pp,qq,:])
-                self.L[:]                += np.cumsum(self.bin_cl['EBo_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EsBd_ij_b'][pp,qq,:])
-                if 'beta' in self.fit:
-                    # Asd*beta
-                    self.DELTA[:]        += np.cumsum(self.bin_cl['EsBd_ij_b'][ii,jj,:]  *self.iC[:,mm,nn]* self.bin_cl['EEcmb_ij_b'][pp,qq,:]) 
-                    self.eta[:]          += np.cumsum(self.bin_cl['EdBs_ij_b'][ii,jj,:]  *self.iC[:,mm,nn]* self.bin_cl['EEcmb_ij_b'][pp,qq,:]) 
-                    self.theta[:]        += np.cumsum(self.bin_cl['EdBs_ij_b'][ii,jj,:]  *self.iC[:,mm,nn]* self.bin_cl['BBcmb_ij_b'][pp,qq,:]) 
-                    self.delta[:]        += np.cumsum(self.bin_cl['EsBd_ij_b'][ii,jj,:]  *self.iC[:,mm,nn]* self.bin_cl['BBcmb_ij_b'][pp,qq,:])
-                if 'Ad' in self.fit:
-                    # Asd*Ad
-                    self.K[:]            += np.cumsum(self.bin_cl['EdBs_ij_b'][ii,jj,:]  *self.iC[:,mm,nn]* self.bin_cl['EBd_ij_b'][pp,qq,:])
-                    self.xi[:]           += np.cumsum(self.bin_cl['EsBd_ij_b'][ii,jj,:]  *self.iC[:,mm,nn]* self.bin_cl['EBd_ij_b'][pp,qq,:])
-                if 'As' in self.fit:
-                    # Asd*As
-                    self.Q[:]            += np.cumsum(self.bin_cl['EBs_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EsBd_ij_b'][pp,qq,:])
-                    self.V[:]            += np.cumsum(self.bin_cl['EBs_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EdBs_ij_b'][pp,qq,:])
-   
-    #TODO PDP: these ones can be further optimised but I'm leaving them like this
-    # for now to debug the new code structure first
-    def __total__(self):
-        for MN_pair in self.mle.MNidx:
-            ii, jj, pp, qq, mm, nn = self.mle.get_index(MN_pair)
-            # common to all fits (basic alpha*alpha fit)
-            self.B_ijpq[ii,jj,pp,qq]    = np.sum(self.bin_cl['BBo_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['BBo_ij_b'][pp,qq,:])
-            self.E_ijpq[ii,jj,pp,qq]    = np.sum(self.bin_cl['EEo_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EEo_ij_b'][pp,qq,:])
-            self.I_ijpq[ii,jj,pp,qq]    = np.sum(self.bin_cl['BBo_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EEo_ij_b'][pp,qq,:]) 
-            self.D_ij[ii,jj]           += np.sum(self.bin_cl['EEo_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EBo_ij_b'][pp,qq,:])
-            self.H_ij[ii,jj]           += np.sum(self.bin_cl['BBo_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EBo_ij_b'][pp,qq,:])
-            self.A                     += np.sum(self.bin_cl['EBo_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EBo_ij_b'][pp,qq,:])
-            # only used in particular combinations 
-            if 'beta' in self.fit:
-                # beta*beta and beta*alpha 
-                self.tau_ij[ii,jj]     += np.sum(self.bin_cl['EEo_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EEcmb_ij_b'][pp,qq,:])
-                self.varphi_ij[ii,jj]  += np.sum(self.bin_cl['EEo_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['BBcmb_ij_b'][pp,qq,:])
-                self.ene_ij[ii,jj]     += np.sum(self.bin_cl['BBo_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EEcmb_ij_b'][pp,qq,:])
-                self.epsilon_ij[ii,jj] += np.sum(self.bin_cl['BBo_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['BBcmb_ij_b'][pp,qq,:])
-                self.C                 += np.sum(self.bin_cl['EEcmb_ij_b'][ii,jj,:] *self.iC[:,mm,nn]* self.bin_cl['BBcmb_ij_b'][pp,qq,:]) 
-                self.F                 += np.sum(self.bin_cl['EEcmb_ij_b'][ii,jj,:] *self.iC[:,mm,nn]* self.bin_cl['EEcmb_ij_b'][pp,qq,:]) 
-                self.G                 += np.sum(self.bin_cl['BBcmb_ij_b'][ii,jj,:] *self.iC[:,mm,nn]* self.bin_cl['BBcmb_ij_b'][pp,qq,:]) 
-                self.O                 += np.sum(self.bin_cl['EBo_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EEcmb_ij_b'][pp,qq,:]) 
-                self.P                 += np.sum(self.bin_cl['EBo_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['BBcmb_ij_b'][pp,qq,:])
-            if 'Ad' in self.fit:
-                # Ad*Ad and Ad*alpha
-                self.sigma_ij[ii,jj]   += np.sum(self.bin_cl['EEo_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EBd_ij_b'][pp,qq,:])
-                self.omega_ij[ii,jj]   += np.sum(self.bin_cl['BBo_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EBd_ij_b'][pp,qq,:])
-                self.R                 += np.sum(self.bin_cl['EBd_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EBd_ij_b'][pp,qq,:])
-                self.N                 += np.sum(self.bin_cl['EBo_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EBd_ij_b'][pp,qq,:])
-                if 'beta' in self.fit:
-                    # Ad*beta
-                    self.LAMBDA        += np.sum(self.bin_cl['EBd_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EEcmb_ij_b'][pp,qq,:])
-                    self.mu            += np.sum(self.bin_cl['EBd_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['BBcmb_ij_b'][pp,qq,:])
-            if 'As' in self.fit:
-                # As*As and As*alpha
-                self.nu_ij[ii,jj]      += np.sum(self.bin_cl['EEo_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EBs_ij_b'][pp,qq,:])
-                self.psi_ij[ii,jj]     += np.sum(self.bin_cl['BBo_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EBs_ij_b'][pp,qq,:]) 
-                self.S                 += np.sum(self.bin_cl['EBs_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EBs_ij_b'][pp,qq,:])
-                self.J                 += np.sum(self.bin_cl['EBo_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EBs_ij_b'][pp,qq,:])
-                if 'beta' in self.fit:
-                    # As*beta
-                    self.X             += np.sum(self.bin_cl['EBs_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EEcmb_ij_b'][pp,qq,:])
-                    self.Y             += np.sum(self.bin_cl['EBs_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['BBcmb_ij_b'][pp,qq,:])
-                if 'Ad' in self.fit:
-                    # As*Ad
-                    self.W             += np.sum(self.bin_cl['EBs_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EBd_ij_b'][pp,qq,:])
-            if 'Asd' in self.fit:
-                # Asd*Asd and Asd*alpha
-                self.pi_ij[ii,jj]      += np.sum(self.bin_cl['EEo_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EsBd_ij_b'][pp,qq,:])
-                self.rho_ij[ii,jj]     += np.sum(self.bin_cl['EEo_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EdBs_ij_b'][pp,qq,:])
-                self.phi_ij[ii,jj]     += np.sum(self.bin_cl['BBo_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EsBd_ij_b'][pp,qq,:])
-                self.OMEGA_ij[ii,jj]   += np.sum(self.bin_cl['BBo_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EdBs_ij_b'][pp,qq,:])
-                self.T                 += np.sum(self.bin_cl['EsBd_ij_b'][ii,jj,:]  *self.iC[:,mm,nn]* self.bin_cl['EsBd_ij_b'][pp,qq,:])
-                self.U                 += np.sum(self.bin_cl['EdBs_ij_b'][ii,jj,:]  *self.iC[:,mm,nn]* self.bin_cl['EdBs_ij_b'][pp,qq,:])
-                self.Z                 += np.sum(self.bin_cl['EsBd_ij_b'][ii,jj,:]  *self.iC[:,mm,nn]* self.bin_cl['EdBs_ij_b'][pp,qq,:])
-                self.M                 += np.sum(self.bin_cl['EBo_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EdBs_ij_b'][pp,qq,:])
-                self.L                 += np.sum(self.bin_cl['EBo_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EsBd_ij_b'][pp,qq,:])
-                if 'beta' in self.fit:
-                    # Asd*beta
-                    self.DELTA         += np.sum(self.bin_cl['EsBd_ij_b'][ii,jj,:]  *self.iC[:,mm,nn]* self.bin_cl['EEcmb_ij_b'][pp,qq,:]) 
-                    self.eta           += np.sum(self.bin_cl['EdBs_ij_b'][ii,jj,:]  *self.iC[:,mm,nn]* self.bin_cl['EEcmb_ij_b'][pp,qq,:]) 
-                    self.theta         += np.sum(self.bin_cl['EdBs_ij_b'][ii,jj,:]  *self.iC[:,mm,nn]* self.bin_cl['BBcmb_ij_b'][pp,qq,:]) 
-                    self.delta         += np.sum(self.bin_cl['EsBd_ij_b'][ii,jj,:]  *self.iC[:,mm,nn]* self.bin_cl['BBcmb_ij_b'][pp,qq,:])
-                if 'Ad' in self.fit:
-                    # Asd*Ad
-                    self.K             += np.sum(self.bin_cl['EdBs_ij_b'][ii,jj,:]  *self.iC[:,mm,nn]* self.bin_cl['EBd_ij_b'][pp,qq,:])
-                    self.xi            += np.sum(self.bin_cl['EsBd_ij_b'][ii,jj,:]  *self.iC[:,mm,nn]* self.bin_cl['EBd_ij_b'][pp,qq,:])
-                if 'As' in self.fit:
-                    # Asd*As
-                    self.Q             += np.sum(self.bin_cl['EBs_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EsBd_ij_b'][pp,qq,:])
-                    self.V             += np.sum(self.bin_cl['EBs_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EdBs_ij_b'][pp,qq,:])
-
-    def __ell__(self):
-        for MN_pair in self.mle.MNidx:
-            ii, jj, pp, qq, mm, nn = self.mle.get_index(MN_pair)
-            # common to all fits (basic alpha*alpha fit)
-            self.B_ijpq[ii,jj,pp,qq,:]    = moving_sum(self.bin_cl['BBo_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['BBo_ij_b'][pp,qq,:], self.window)
-            self.E_ijpq[ii,jj,pp,qq,:]    = moving_sum(self.bin_cl['EEo_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EEo_ij_b'][pp,qq,:], self.window)
-            self.I_ijpq[ii,jj,pp,qq,:]    = moving_sum(self.bin_cl['BBo_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EEo_ij_b'][pp,qq,:], self.window) 
-            self.D_ij[ii,jj,:]           += moving_sum(self.bin_cl['EEo_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EBo_ij_b'][pp,qq,:], self.window)
-            self.H_ij[ii,jj,:]           += moving_sum(self.bin_cl['BBo_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EBo_ij_b'][pp,qq,:], self.window)
-            self.A[:]                    += moving_sum(self.bin_cl['EBo_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EBo_ij_b'][pp,qq,:], self.window)
-            # only used in particular combinations
-            if 'beta' in self.fit:
-                # beta*beta and beta*alpha 
-                self.tau_ij[ii,jj,:]     += moving_sum(self.bin_cl['EEo_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EEcmb_ij_b'][pp,qq,:], self.window)
-                self.varphi_ij[ii,jj,:]  += moving_sum(self.bin_cl['EEo_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['BBcmb_ij_b'][pp,qq,:], self.window)
-                self.ene_ij[ii,jj,:]     += moving_sum(self.bin_cl['BBo_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EEcmb_ij_b'][pp,qq,:], self.window)
-                self.epsilon_ij[ii,jj,:] += moving_sum(self.bin_cl['BBo_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['BBcmb_ij_b'][pp,qq,:], self.window)
-                self.C[:]                += moving_sum(self.bin_cl['EEcmb_ij_b'][ii,jj,:] *self.iC[:,mm,nn]* self.bin_cl['BBcmb_ij_b'][pp,qq,:], self.window) 
-                self.F[:]                += moving_sum(self.bin_cl['EEcmb_ij_b'][ii,jj,:] *self.iC[:,mm,nn]* self.bin_cl['EEcmb_ij_b'][pp,qq,:], self.window) 
-                self.G[:]                += moving_sum(self.bin_cl['BBcmb_ij_b'][ii,jj,:] *self.iC[:,mm,nn]* self.bin_cl['BBcmb_ij_b'][pp,qq,:], self.window) 
-                self.O[:]                += moving_sum(self.bin_cl['EBo_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EEcmb_ij_b'][pp,qq,:], self.window) 
-                self.P[:]                += moving_sum(self.bin_cl['EBo_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['BBcmb_ij_b'][pp,qq,:], self.window)
-            if 'Ad' in self.fit:
-                # Ad*Ad and Ad*alpha
-                self.sigma_ij[ii,jj,:]   += moving_sum(self.bin_cl['EEo_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EBd_ij_b'][pp,qq,:], self.window)
-                self.omega_ij[ii,jj,:]   += moving_sum(self.bin_cl['BBo_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EBd_ij_b'][pp,qq,:], self.window)
-                self.R[:]                += moving_sum(self.bin_cl['EBd_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EBd_ij_b'][pp,qq,:], self.window)
-                self.N[:]                += moving_sum(self.bin_cl['EBo_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EBd_ij_b'][pp,qq,:], self.window)
-                if 'beta' in self.fit:
-                    # Ad*beta
-                    self.LAMBDA[:]       += moving_sum(self.bin_cl['EBd_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EEcmb_ij_b'][pp,qq,:], self.window)
-                    self.mu[:]           += moving_sum(self.bin_cl['EBd_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['BBcmb_ij_b'][pp,qq,:], self.window)
-            if 'As' in self.fit:
-                # As*As and As*alpha
-                self.nu_ij[ii,jj,:]      += moving_sum(self.bin_cl['EEo_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EBs_ij_b'][pp,qq,:], self.window)
-                self.psi_ij[ii,jj,:]     += moving_sum(self.bin_cl['BBo_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EBs_ij_b'][pp,qq,:], self.window) 
-                self.S[:]                += moving_sum(self.bin_cl['EBs_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EBs_ij_b'][pp,qq,:], self.window)
-                self.J[:]                += moving_sum(self.bin_cl['EBo_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EBs_ij_b'][pp,qq,:], self.window)
-                if 'beta' in self.fit:
-                    # As*beta
-                    self.X[:]            += moving_sum(self.bin_cl['EBs_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EEcmb_ij_b'][pp,qq,:], self.window)
-                    self.Y[:]            += moving_sum(self.bin_cl['EBs_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['BBcmb_ij_b'][pp,qq,:], self.window)
-                if 'Ad' in self.fit:
-                    # As*Ad
-                    self.W[:]            += moving_sum(self.bin_cl['EBs_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EBd_ij_b'][pp,qq,:], self.window)
-            if 'Asd' in self.fit:
-                # Asd*Asd and Asd*alpha
-                self.pi_ij[ii,jj,:]      += moving_sum(self.bin_cl['EEo_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EsBd_ij_b'][pp,qq,:], self.window)
-                self.rho_ij[ii,jj,:]     += moving_sum(self.bin_cl['EEo_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EdBs_ij_b'][pp,qq,:], self.window)
-                self.phi_ij[ii,jj,:]     += moving_sum(self.bin_cl['BBo_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EsBd_ij_b'][pp,qq,:], self.window)
-                self.OMEGA_ij[ii,jj,:]   += moving_sum(self.bin_cl['BBo_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EdBs_ij_b'][pp,qq,:], self.window)
-                self.T[:]                += moving_sum(self.bin_cl['EsBd_ij_b'][ii,jj,:]  *self.iC[:,mm,nn]* self.bin_cl['EsBd_ij_b'][pp,qq,:], self.window)
-                self.U[:]                += moving_sum(self.bin_cl['EdBs_ij_b'][ii,jj,:]  *self.iC[:,mm,nn]* self.bin_cl['EdBs_ij_b'][pp,qq,:], self.window)
-                self.Z[:]                += moving_sum(self.bin_cl['EsBd_ij_b'][ii,jj,:]  *self.iC[:,mm,nn]* self.bin_cl['EdBs_ij_b'][pp,qq,:], self.window)
-                self.M[:]                += moving_sum(self.bin_cl['EBo_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EdBs_ij_b'][pp,qq,:], self.window)
-                self.L[:]                += moving_sum(self.bin_cl['EBo_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EsBd_ij_b'][pp,qq,:], self.window)
-                if 'beta' in self.fit:
-                    # Asd*beta
-                    self.DELTA[:]        += moving_sum(self.bin_cl['EsBd_ij_b'][ii,jj,:]  *self.iC[:,mm,nn]* self.bin_cl['EEcmb_ij_b'][pp,qq,:], self.window) 
-                    self.eta[:]          += moving_sum(self.bin_cl['EdBs_ij_b'][ii,jj,:]  *self.iC[:,mm,nn]* self.bin_cl['EEcmb_ij_b'][pp,qq,:], self.window) 
-                    self.theta[:]        += moving_sum(self.bin_cl['EdBs_ij_b'][ii,jj,:]  *self.iC[:,mm,nn]* self.bin_cl['BBcmb_ij_b'][pp,qq,:], self.window) 
-                    self.delta[:]        += moving_sum(self.bin_cl['EsBd_ij_b'][ii,jj,:]  *self.iC[:,mm,nn]* self.bin_cl['BBcmb_ij_b'][pp,qq,:], self.window)
-                if 'Ad' in self.fit:
-                    # Asd*Ad
-                    self.K[:]            += moving_sum(self.bin_cl['EdBs_ij_b'][ii,jj,:]  *self.iC[:,mm,nn]* self.bin_cl['EBd_ij_b'][pp,qq,:], self.window)
-                    self.xi[:]           += moving_sum(self.bin_cl['EsBd_ij_b'][ii,jj,:]  *self.iC[:,mm,nn]* self.bin_cl['EBd_ij_b'][pp,qq,:], self.window)
-                if 'As' in self.fit:
-                    # Asd*As
-                    self.Q[:]            += moving_sum(self.bin_cl['EBs_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EsBd_ij_b'][pp,qq,:], self.window)
-                    self.V[:]            += moving_sum(self.bin_cl['EBs_ij_b'][ii,jj,:]   *self.iC[:,mm,nn]* self.bin_cl['EdBs_ij_b'][pp,qq,:], self.window)
-
-
-
-########################################################################       
+     
 
 class MLE:
     fit_options = ["alpha", "Ad + alpha", "As + Ad + alpha", "As + Asd + Ad + alpha",
